doc: added docs for scalars
