+++ docs/codegen/scalars/custom_scalars.md
@@ -1,0 +1,7 @@
+
+# Custom scalars
+**QtGQL** provides some custom scalars by default, they are intended to be comatible with
+[Strawberry-GraphQL](https://strawberry.rocks/docs/types/scalars#scalars) custom scalars,
+though you can override them.
+
+::: qtgql.codegen.py.custom_scalars
