+++ tests/test_sample_ui/__temp.py
@@ -5,7 +5,7 @@ from typing import Optional
 from PySide6.QtCore import QObject, Signal
 from qtgql import qproperty
 from qtgql.codegen.py.bases import BaseGraphQLObject, BaseModel
-from qtgql.codegen.py.scalars import DateScalar, DateTimeScalar, DecimalScalar
+from qtgql.codegen.py.custom_scalars import DateScalar, DateTimeScalar, DecimalScalar
 
 
 class SCALARS:
